+++ lib/fb_news_feed.dart
@@ -3,6 +3,8 @@ import 'dart:convert';
 
 // Flutter imports:
 import 'package:flutter/material.dart';
+import 'package:flutter_fb_news/fb_news_attachments_share.dart';
+import 'package:flutter_fb_news/fb_news_attachments_video_direct_response.dart';
 
 // Project imports:
 import 'fb_news_attachments_photos.dart';
@@ -25,9 +27,9 @@ class FbNewsFeed extends StatelessWidget {
   final FbNewsConfig config;
 
   const FbNewsFeed({
-    @required this.feedResponse,
-    @required this.profilePictureUrl,
-    @required this.config,
+    required this.feedResponse,
+    required this.profilePictureUrl,
+    required this.config,
   });
 
   @override
@@ -69,14 +71,14 @@ class FbNewsFeedLayout extends StatelessWidget {
   /// Only for internal use of flutter_fb_news
   final String profilePictureUrl;
   const FbNewsFeedLayout({
-    @required this.feed,
-    @required this.profilePictureUrl,
-    @required this.config,
+    required this.feed,
+    required this.profilePictureUrl,
+    required this.config,
   });
 
-  bool hasField(String internalKey) {
+  bool hasField(FbNewsFieldName key) {
     return config.fields
-        .where((element) => element.internalKey == internalKey)
+        .where((element) => element.internalKey == key.internalKey)
         .isNotEmpty;
   }
 
@@ -90,7 +92,7 @@ class FbNewsFeedLayout extends StatelessWidget {
         child: Column(
           children: [
             hasField(
-              FbNewsFields.header.internalKey,
+              FbNewsFields.header,
             )
                 ? FbNewsHeader(
                     feed: feed,
@@ -99,7 +101,7 @@ class FbNewsFeedLayout extends StatelessWidget {
                   )
                 : Container(),
             hasField(
-              FbNewsFields.message.internalKey,
+              FbNewsFields.message,
             )
                 ? FbNewsMessage(
                     feed: feed,
@@ -107,27 +109,35 @@ class FbNewsFeedLayout extends StatelessWidget {
                   )
                 : Container(),
             hasField(
-                      "attachments",
-                    ) ||
-                    hasField(
-                      FbNewsFields.attachmentsPhotos.internalKey,
-                    )
+              FbNewsFields.attachmentsPhotos,
+            )
                 ? FbNewsAttachmentsPhotos(
                     feed: feed,
                   )
                 : Container(),
             hasField(
-                      "attachments",
-                    ) ||
-                    hasField(
-                      FbNewsFields.attachmentsVideos.internalKey,
-                    )
+              FbNewsFields.attachmentsVideos,
+            )
                 ? FbNewsAttachmentsVideos(
                     feed: feed,
                   )
                 : Container(),
             hasField(
-              FbNewsFields.footer.internalKey,
+              FbNewsFields.attachmentsShare,
+            )
+                ? FbNewsAttachmentsShares(
+                    feed: feed,
+                  )
+                : Container(),
+            hasField(
+              FbNewsFields.attachmentsVideoDirectResponseAutoplay,
+            )
+                ? FbNewsAttachmentsVideoDirectResponse(
+                    feed: feed,
+                  )
+                : Container(),
+            hasField(
+              FbNewsFields.footer,
             )
                 ? FbNewsFooter(
                     feed: feed,