+++ test/TestMaterialWidget.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 class TestMaterialWidget extends StatelessWidget {
   final Widget child;
 
-  const TestMaterialWidget({Key key, this.child}) : super(key: key);
+  const TestMaterialWidget({Key? key, required this.child}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {